refactor(capture): the pre-store gates are one helper
