+++ backend/internal/modules/capture/sinkexclusion.go
@@ -10,6 +10,7 @@ import (
 	"github.com/jackc/pgx/v5"
 
 	"github.com/gradionhq/margince/backend/internal/shared/kernel/ids"
+	"github.com/gradionhq/margince/backend/internal/shared/kernel/pipelinetrace"
 	"github.com/gradionhq/margince/backend/internal/shared/ports/connector"
 )
 
@@ -22,6 +23,46 @@ const (
 	reasonExcludedDomain  = "excluded_domain"
 	actionCaptureExcluded = "capture_excluded"
 )
+
+// dropBeforeStoreTx runs the gates that keep a message out of the CRM
+// entirely — the colleagues-only gate and the exclusion lists — and answers
+// the sentence the connector's skip carries, or "" when the message may be
+// stored. Each drop leaves its breadcrumb (the operator's record) and its
+// trace (the member's answer to "why did this never appear"), and neither
+// names an address.
+func (s *Sink) dropBeforeStoreTx(ctx context.Context, tx pgx.Tx, rec connector.NormalizedRecord) (string, error) {
+	// Colleague correspondence is not evidence of a customer relationship — it
+	// is two employees talking, and the CRM was never asked to hold it.
+	internal, err := s.internalOnlyTx(ctx, tx, rec)
+	if err != nil {
+		return "", err
+	}
+	if internal {
+		if err := s.logBreadcrumbTx(ctx, tx, actionCaptureInternalDropped, rec, reasonInternalOnly); err != nil {
+			return "", err
+		}
+		if err := s.traceTx(ctx, tx, rec, pipelinetrace.StageInternalDrop, TraceInternal, reasonInternalOnly); err != nil {
+			return "", err
+		}
+		return "all participants are on the workspace's own domains", nil
+	}
+	// A message the workspace or this mailbox's owner ruled out is not the
+	// CRM's to hold either.
+	excluded, err := excludedTx(ctx, tx, rec)
+	if err != nil {
+		return "", err
+	}
+	if excluded == "" {
+		return "", nil
+	}
+	if err := s.logBreadcrumbTx(ctx, tx, actionCaptureExcluded, rec, excluded); err != nil {
+		return "", err
+	}
+	if err := s.traceTx(ctx, tx, rec, pipelinetrace.StageInternalDrop, TraceSuppressed, excluded); err != nil {
+		return "", err
+	}
+	return "a capture exclusion rule keeps this message out (" + excluded + ")", nil
+}
 
 // excludedTx reports whether any address this message names — sender,
 // recipients, copies — is kept out by a workspace exclusion, or by one the
